Update Lesson_3_Let_Users_Mint_Your_NFT.md
Proposed change to reflect fix to a bug in the code for the "Mint an NFT" button. The bug was allowing you to pass to the member page even if you rejected the NFT mint.  This change is from pull request by another user (#mintNFT Bugfix #2), however, there is a syntax error in that PR -- a ";'" immediately before .catch((err) -- that prevented replit from compiling. The change here removes the ";".